Fix InBag iterator bug
Before this commit, InBag iterator's Pop() will get value and advance to the
next item. The problem of that is the next item may not exist! We fix the bug
by using the information in InBag iterator: we know exactly how many items
we need, so we just do not advance if we are not sure if the next item exist